+++ buession-core/src/main/java/com/buession/core/validator/constraintvalidators/PortConstraintValidator.java
@@ -19,7 +19,7 @@
  * +-------------------------------------------------------------------------------------------------------+
  * | License: http://www.apache.org/licenses/LICENSE-2.0.txt 										       |
  * | Author: Yong.Teng <webmaster@buession.com> 													       |
- * | Copyright @ 2013-2022 Buession.com Inc.														       |
+ * | Copyright @ 2013-2023 Buession.com Inc.														       |
  * +-------------------------------------------------------------------------------------------------------+
  */
 package com.buession.core.validator.constraintvalidators;
@@ -35,10 +35,6 @@ import javax.validation.ConstraintValidatorContext;
  * @since 2.0.0
  */
 public abstract class PortConstraintValidator implements ConstraintValidator<Port, Integer> {
-
-	@Override
-	public void initialize(Port port){
-	}
 
 	@Override
 	public boolean isValid(Integer value, ConstraintValidatorContext context){