+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 - **buession-core：** StringUtils 增加 random 增加支持字符串
 - **buession-geoip：** 增加返回 autonomous system number、autonomous system organization
 - **buession-geoip：** 增加可支持设置 asn 库地址或流
+- **buession-httpclient：** 支持原生 RequestBody 子类转换
 - **buession-web：** 增加 web binder 转换器工厂 IgnoreCaseEnumConverterFactory 忽略大小写将字符串转换为枚举值
 - **buession-web：** 获取客户端真实 IP，增加标头 Client-IP 的支持